+++ Shared/ViewModels/DaneDoPlanowaniaViewModel.cs
@@ -13,7 +13,7 @@ namespace GEORGE.Shared.ViewModels
         public ZleceniaProdukcyjneDto? ZleceniaProdukcyjneDto { get; set; }
         public string? Wyrob { get; set; }
         public int DomyslnyCzasProdukcji{ get; set; } = 24;//  Wartość domyślna
-        public int JednostkiNaZlecenie { get; set; }
+        public float JednostkiNaZlecenie { get; set; }
         public string? NumerZlecenia { get; set; }
         public string? RowIdLiniiProdukcyjnej { get; set; }
         public bool ZlecenieWewnetrzne { get; set; } = false;
